Positioners cleanup
Use qreal for positions and spacing instead of ints, and make the
positionedItems list take care of filtering out invisible items.

+++ src/quick/items/qquickpositioners_p_p.h
@@ -84,7 +84,7 @@ public:
         childrenDoNotOverlap = true;
     }
 
-    int spacing;
+    qreal spacing;
 
     QQuickBasePositioner::PositionerType type;
     QDeclarativeTransition *moveTransition;
